Adding paging via Jetpack Paging

+++ news-data/src/main/java/dev/androidbroadcast/news/data/model/Article.kt
@@ -3,7 +3,7 @@ package dev.androidbroadcast.news.data.model
 import java.util.Date
 
 public data class Article(
-    val cacheId: Long = ID_NONE,
+    val cacheId: Int = ID_NONE,
     val source: Source,
     val author: String?,
     val title: String,
@@ -17,7 +17,7 @@ public data class Article(
         /**
          * Специальный ID для обозначения что ID нету
          */
-        public const val ID_NONE: Long = 0L
+        public const val ID_NONE: Int = 0
     }
 }
